Moved ktest utilities to ktest.c
